pyproject.toml
[project]
name = "mkinf"
version = "0.1.8"
description = "mkinf python SDK"
authors = [{ name = "mkinf", email = "info@mkinf.io" }]
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "requests>=2.32.3,<3",
    "python-dotenv>=1.0.1,<2",
    "langchain-core>=0.3.31,<0.4",
    "pydantic>=2.10.5,<3",
    "toml>=0.10.2",
]

[dependency-groups]
dev = [
  "langchain>=0.3.9",
  "langchain-community>=0.3.9",
  "langchain-openai>=0.2.10",
  "pydantic>=2.10.2",
  "langgraph>=0.2.53",
  "streamlit>=1.40.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
